+++ src/Quoter.php
@@ -6,24 +6,23 @@ namespace Yiisoft\Db\Mysql;
 
 use Yiisoft\Db\Schema\Quoter as BaseQuoter;
 
-use function is_string;
-use function str_replace;
+use function strtr;
 
 /**
  * Implements MySQL, MariaDB quoting and unquoting methods.
  */
 final class Quoter extends BaseQuoter
 {
-    public function quoteValue(mixed $value): mixed
+    public function quoteValue(string $value): string
     {
-        if (!is_string($value)) {
-            return $value;
-        }
-
-        return "'" . str_replace(
-            ['\\', "\x00", "\n", "\r", "'", '"', "\x1a"],
-            ['\\\\', '\\0', '\\n', '\\r', "\'", '\"', '\\Z'],
-            $value
-        ) . "'";
+        return "'" . strtr($value, [
+            '\\' => '\\\\',
+            "\x00" => '\\0',
+            "\n" => '\\n',
+            "\r" => '\\r',
+            "'" => "\'",
+            '"' => '\"',
+            "\x1a" => '\\Z',
+        ]) . "'";
     }
 }